# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api bf_immo",
        "description": "test",
        "contact": {
            "email": "florenias@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/property/all": {
            "get": {
                "tags": [
                    "All Properties"
                ],
                "summary": "Get List Of Properties",
                "description": "Returns all properties and associated attributes.",
                "operationId": "getAllProperties",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/property/detail/{id}": {
            "get": {
                "tags": [
                    "All Properties"
                ],
                "summary": "Get List Of One Property",
                "description": "Returns one property at a time and its associated attributes.",
                "operationId": "getOneProperty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/property/delete/{id}": {
            "delete": {
                "tags": [
                    "All Properties"
                ],
                "summary": "Delete a single property",
                "description": "delete a single property based on the ID supplied",
                "operationId": "deleteProperty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of property to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "property deleted"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}